door content separated, clean up
